Added several variable blocks
Added blocks to create and delete variables. I also added a block that
returns a variable object from a string with the name of the variable.
While this may not be useful right now, I can see how it'd be useful if
variable blocks could accept them. Lastly, I added a block that returns
the value of a variable given as a string. This allows you to set the
value of variables created in the project. Like I said before, these
blocks are still a bit buggy, but still pretty useful.